new endpoint and websocket with json to build insights dashboards

+++ backend-django/api_layer/views.py
@@ -1,5 +1,3 @@
-from asgiref.sync import async_to_sync
-from channels.layers import get_channel_layer
 from django.contrib.auth import authenticate
 from django.utils import timezone
 from rest_framework import status
@@ -9,6 +7,8 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework.generics import CreateAPIView
 
+from ingredients.models import Ingredient
+from .consumers import update_websocket_dashboard, message_websocket_dashboard
 from .serializers import UserSerializer, IngredientSerializer
 
 
@@ -32,11 +32,12 @@ class LogoutView(APIView):
     authentication_classes = [TokenAuthentication]
 
     def post(self, request):
-        request.user.auth_token.delete()
+        if request.user.is_authenticated:
+            request.user.auth_token.delete()
         return Response(status=204)
 
 
-class IngredientProduceView(APIView):
+class ProduceIngredientView(APIView):
     authentication_classes = (TokenAuthentication,)
 
     def post(self, request):
@@ -47,12 +48,35 @@ class IngredientProduceView(APIView):
         serializer = IngredientSerializer(data=data)
         if serializer.is_valid():
             ingredient = serializer.save()
-            channel_layer = get_channel_layer()
-            message = {
-                'type': 'chat_message',
-                'message': f'{ingredient}'
-            }
-            async_to_sync(channel_layer.group_send)('broadcast', message)
+            update_websocket_dashboard()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+
+class BuyIngredientView(APIView):
+    authentication_classes = (TokenAuthentication,)
+
+    def post(self, request):
+        data = request.data
+
+        chosen_ingredient = Ingredient.objects.filter(
+            buyer__isnull=True,
+            bought_at__isnull=True,
+            name__iexact=data.get('name'),
+        ).exclude(
+            producer=request.user
+        ).order_by('?').first()
+
+        if not chosen_ingredient:
+            message_websocket_dashboard(data.get('name'))
+
+        data['id'] = chosen_ingredient.id
+        data['buyer'] = request.user.id
+        data['bought_at'] = timezone.now()
+
+        serializer = IngredientSerializer(instance=chosen_ingredient, data=data, partial=True)
+        if serializer.is_valid():
+            ingredient = serializer.save()
+            update_websocket_dashboard()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
